requirements.txt
Django>=5.2,<5.3
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
sqlalchemy>=2.0,<2.1
pdfplumber>=0.11,<0.12
python-multipart>=0.0.9,<0.0.10
pydantic>=2.8,<3
python-dotenv>=1.0,<2
scikit-learn>=1.5,<1.6
joblib>=1.4,<2
pytest>=8.3,<9
httpx>=0.27,<0.28
